+++ src/tm_devices/commands/gen_fn2qbf_msodpo/errordetector.py
@@ -122,7 +122,7 @@ Commands and Queries:
     - ERRORDetector:SCRAMBLED?
     - ERRORDetector:SENDEMAIL {OFF|ON}
     - ERRORDetector:SENDEMAIL?
-    - ERRORDetector:SIGnaltype {SATAGEN<x>|PRBS7|PRBS11|USB3|PCIEGEN<x>|PRBS23|CUSTOM|ANY8B10B|PRBS16|PRBS9}
+    - ERRORDetector:SIGnaltype {PRBS16|PCIEGEN<x>|PRBS7|ANY8B10B|USB3|PRBS23|CUSTOM|PRBS9|SATAGEN<x>|PRBS11}
     - ERRORDetector:SIGnaltype?
     - ERRORDetector:SSC {ON|OFF}
     - ERRORDetector:SSC?
@@ -1391,7 +1391,7 @@ class ErrordetectorSignaltype(SCPICmdWrite, SCPICmdRead):
 
     SCPI Syntax:
         ```
-        - ERRORDetector:SIGnaltype {SATAGEN<x>|PRBS7|PRBS11|USB3|PCIEGEN<x>|PRBS23|CUSTOM|ANY8B10B|PRBS16|PRBS9}
+        - ERRORDetector:SIGnaltype {PRBS16|PCIEGEN<x>|PRBS7|ANY8B10B|USB3|PRBS23|CUSTOM|PRBS9|SATAGEN<x>|PRBS11}
         - ERRORDetector:SIGnaltype?
         ```
 
@@ -5654,7 +5654,7 @@ class Errordetector(SCPICmdRead):
 
         SCPI Syntax:
             ```
-            - ERRORDetector:SIGnaltype {SATAGEN<x>|PRBS7|PRBS11|USB3|PCIEGEN<x>|PRBS23|CUSTOM|ANY8B10B|PRBS16|PRBS9}
+            - ERRORDetector:SIGnaltype {PRBS16|PCIEGEN<x>|PRBS7|ANY8B10B|USB3|PRBS23|CUSTOM|PRBS9|SATAGEN<x>|PRBS11}
             - ERRORDetector:SIGnaltype?
             ```