+++ src/tm_devices/commands/mso2_commands.py
@@ -23,8 +23,6 @@ from .gen_1zn03_mso.diag import Diag
 from .gen_1zn03_mso.display import Display
 from .gen_1zn03_mso.fpanel import Fpanel
 from .gen_1zn03_mso.horizontal import Horizontal
-from .gen_1zn03_mso.lic import Lic
-from .gen_1zn03_mso.license import License
 from .gen_1zn03_mso.mask import Mask
 from .gen_1zn03_mso.math import Math
 from .gen_1zn03_mso.measurement import Measurement
@@ -39,57 +37,68 @@ from .gen_1zn03_mso.search import Search
 from .gen_1zn03_mso.select import Select
 from .gen_1zn03_mso.touchscreen import Touchscreen
 from .gen_1zn03_mso.trigger import Trigger
+from .gen_c69az_msotekscopepc.lic import Lic
+from .gen_c69az_msotekscopepc.license import License
 from .gen_e3h2zs_lpdmso.afg import Afg
-from .gen_e3h2zs_lpdmso.autosavepitimeout import Autosavepitimeout
-from .gen_e3h2zs_lpdmso.autosaveuitimeout import Autosaveuitimeout
 from .gen_e3h2zs_lpdmso.autoset import Autoset
-from .gen_e3h2zs_lpdmso.bustable import Bustable
 from .gen_e3h2zs_lpdmso.calibrate import Calibrate
-from .gen_e3h2zs_lpdmso.configuration import Configuration
 from .gen_e3h2zs_lpdmso.connected import Connected
-from .gen_e3h2zs_lpdmso.curve import Curve
-from .gen_e3h2zs_lpdmso.curvestream import Curvestream
-from .gen_e3h2zs_lpdmso.customtable import Customtable
-from .gen_e3h2zs_lpdmso.date import Date
 from .gen_e3h2zs_lpdmso.ethernet import Ethernet
-from .gen_e3h2zs_lpdmso.filesystem import Filesystem
-from .gen_e3h2zs_lpdmso.mainwindow import Mainwindow
-from .gen_e3h2zs_lpdmso.meastable import Meastable
-from .gen_e3h2zs_lpdmso.recall import Recall
-from .gen_e3h2zs_lpdmso.socketserver import Socketserver
-from .gen_e3h2zs_lpdmso.time import Time
-from .gen_e3h2zs_lpdmso.undo import Undo
 from .gen_e3h2zs_lpdmso.usbdevice import Usbdevice
-from .gen_e3h2zs_lpdmso.vertical import Vertical
-from .gen_e3h2zs_lpdmso.wfmoutpre import Wfmoutpre
 from .gen_e4de2d_lpdmsomdo.clear import Clear
-from .gen_e6lgg1_lpdmsodpomdo.totaluptime import Totaluptime
-from .gen_e6606z_lpdmsomdodpo.pause import Pause
+from .gen_e6bmgw_lpdmsotekscopepcdpomdo.totaluptime import Totaluptime
+from .gen_e6wozn_lpdmsotekscopepcmdodpo.pause import Pause
+from .gen_e47rsg_lpdmsotekscopepc.autosavepitimeout import Autosavepitimeout
+from .gen_e47rsg_lpdmsotekscopepc.autosaveuitimeout import Autosaveuitimeout
+from .gen_e47rsg_lpdmsotekscopepc.bustable import Bustable
+from .gen_e47rsg_lpdmsotekscopepc.configuration import Configuration
+from .gen_e47rsg_lpdmsotekscopepc.curve import Curve
+from .gen_e47rsg_lpdmsotekscopepc.curvestream import Curvestream
+from .gen_e47rsg_lpdmsotekscopepc.customtable import Customtable
+from .gen_e47rsg_lpdmsotekscopepc.date import Date
+from .gen_e47rsg_lpdmsotekscopepc.filesystem import Filesystem
+from .gen_e47rsg_lpdmsotekscopepc.mainwindow import Mainwindow
+from .gen_e47rsg_lpdmsotekscopepc.meastable import Meastable
+from .gen_e47rsg_lpdmsotekscopepc.recall import Recall
+from .gen_e47rsg_lpdmsotekscopepc.socketserver import Socketserver
+from .gen_e47rsg_lpdmsotekscopepc.time import Time
+from .gen_e47rsg_lpdmsotekscopepc.undo import Undo
+from .gen_e47rsg_lpdmsotekscopepc.vertical import Vertical
+from .gen_e47rsg_lpdmsotekscopepc.wfmoutpre import Wfmoutpre
+from .gen_fsksdy_lpdmsotekscopepcdpomdoafgawgdsa.miscellaneous import Idn, Tst
+from .gen_fsksdy_lpdmsotekscopepcdpomdoafgawgdsa.status_and_error import (
+    Cls,
+    Esr,
+    Opc,
+    Rst,
+    Stb,
+    Wai,
+)
+from .gen_fst7sp_lpdmsotekscopepcmdodpoafgawgdsa.status_and_error import Opt
 from .gen_ft5uww_lpdmsodpomdoafgawgdsa.calibration import Cal
-from .gen_ft5uww_lpdmsodpomdoafgawgdsa.miscellaneous import Idn, Trg, Tst
-from .gen_ft5uww_lpdmsodpomdoafgawgdsa.status_and_error import Cls, Esr, Opc, Rst, Stb, Wai
-from .gen_fteabn_lpdmsomdodpoafgawgdsa.status_and_error import Opt
-from .gen_fug7nl_lpdmsodpomdoawgdsa.status_and_error import Ese, Sre
-from .gen_fuzvln_lpdmsodpodsa.alias import Alias
-from .gen_fuzvln_lpdmsodpodsa.status_and_error import Psc
-from .gen_fx54ua_lpdmsodpomdodsa.allev import Allev
-from .gen_fx54ua_lpdmsodpomdodsa.busy import Busy
-from .gen_fx54ua_lpdmsodpomdodsa.dese import Dese
-from .gen_fx54ua_lpdmsodpomdodsa.event import Event
-from .gen_fx54ua_lpdmsodpomdodsa.evmsg import Evmsg
-from .gen_fx54ua_lpdmsodpomdodsa.evqty import Evqty
-from .gen_fx54ua_lpdmsodpomdodsa.factory import Factory
-from .gen_fx54ua_lpdmsodpomdodsa.header import Header
-from .gen_fx54ua_lpdmsodpomdodsa.id import Id
-from .gen_fx54ua_lpdmsodpomdodsa.miscellaneous import Ddt, Lrn
+from .gen_ft5uww_lpdmsodpomdoafgawgdsa.miscellaneous import Trg
+from .gen_fu6dog_lpdmsotekscopepcdpomdoawgdsa.status_and_error import Ese, Sre
+from .gen_fuq1mi_lpdmsotekscopepcdpodsa.alias import Alias
+from .gen_fuq1mi_lpdmsotekscopepcdpodsa.status_and_error import Psc
+from .gen_fx54ua_lpdmsodpomdodsa.miscellaneous import Ddt
 from .gen_fx54ua_lpdmsodpomdodsa.newpass import Newpass
 from .gen_fx54ua_lpdmsodpomdodsa.password import Password
-from .gen_fx54ua_lpdmsodpomdodsa.rem import Rem
-from .gen_fx54ua_lpdmsodpomdodsa.set import Set
-from .gen_fx54ua_lpdmsodpomdodsa.status_and_error import Pud
 from .gen_fx54ua_lpdmsodpomdodsa.teksecure import Teksecure
-from .gen_fx54ua_lpdmsodpomdodsa.verbose import Verbose
-from .gen_fx54ua_lpdmsodpomdodsa.wavfrm import Wavfrm
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.allev import Allev
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.busy import Busy
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.dese import Dese
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.event import Event
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.evmsg import Evmsg
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.evqty import Evqty
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.factory import Factory
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.header import Header
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.id import Id
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.miscellaneous import Lrn
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.rem import Rem
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.set import Set
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.status_and_error import Pud
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.verbose import Verbose
+from .gen_fxvtmy_lpdmsotekscopepcdpomdodsa.wavfrm import Wavfrm
 from .gen_fzn174_lpdmsodpomdodsa.lock import Lock
 from .gen_fzn174_lpdmsodpomdodsa.unlock import Unlock
 from .helpers import DefaultDictPassKeyToFactory