+++ tests/test_unsupported_device_type.py
@@ -49,6 +49,8 @@ def test_unsupported_device_type_class(device_manager: DeviceManager) -> None:
             "Not all functionality will be available in the device driver. "
             "Please consider contributing to tm_devices to implement official support for "
             "this device type.",
+        ), pytest.warns(
+            UserWarning, match='The "UNSUPPORTED" model is not supported by tm_devices'
         ):
             unsupported_device: CustomUnsupportedDeviceUnitTestOnly = (
                 device_manager.add_unsupported_device(
@@ -72,6 +74,8 @@ def test_unsupported_device_type_class(device_manager: DeviceManager) -> None:
             "Not all functionality will be available in the device driver. "
             "Please consider contributing to tm_devices to implement official support for "
             "this device type.",
+        ), pytest.warns(
+            UserWarning, match='The "UNSUPPORTED" model is not supported by tm_devices'
         ):
             device_manager.load_config_file(
                 Path(__file__).parent / "samples/unsupported_device_type_config.yaml"