+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Things to be included in the next release go here.
 ### Added
 
 - Added curve query support for the MSO2KB series scopes
+- Full Python API support for TekScopePC device.
 
 ---